requirements.txt
fastapi==0.115.0
langchain==0.3.0
langchain_community==0.3.0
langchain_core==0.3.5
langchain_groq==0.2.0
langchain_text_splitters==0.3.0
langchain_experimental==0.3.0
langchainhub
langchain-chroma 
pydantic==2.9.2
python-dotenv==1.0.1
PyYAML==6.0.2
Requests==2.32.3
streamlit==1.38.0
youtube_transcript_api==0.6.2
gpt4all==2.8.2
